+++ sample-compose/src/main/java/com/airbnb/mvrx/compose/sample/ComposeSampleApplication.kt
@@ -1,0 +1,12 @@
+package com.airbnb.mvrx.compose.sample
+
+import android.app.Application
+import com.airbnb.mvrx.Mavericks
+
+class ComposeSampleApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        Mavericks.initialize(this)
+    }
+}