+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
 
 ## 2.1.0
-- Replaced ViewModelContext Activity type from FragmentActivity to its parent class ComponentActivity for better interop with Compose.
+
+### Breaking Changes
+- ActivityViewModelContext and MavericksViewModelFactory now uses ComponentActivity instead of FragmentActivity to improve Compose interop. ComponentActivity is the super class of FragmentActivity so you may need to replace FragmentActivity with ComponentActivity if you using ActivityViewModelContext.
 
 ## Version 2.0.0
 Mavericks 2.0 is a ground up rewrite for coroutines. Check out the [documentation for 2.0](https://airbnb.io/mavericks/#/new-2x) to find out what is new and how to upgrade.